Put source roots back on source root stripped sources.
