+++ src/python/pants/backend/python/rules/coverage.py
@@ -2,11 +2,11 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 import configparser
+import itertools
 import json
 from dataclasses import dataclass
 from io import StringIO
 from typing import Dict, Tuple
-import itertools
 
 import pkg_resources
 
@@ -18,10 +18,12 @@ from pants.backend.python.rules.pex import (
   PexRequirements,
 )
 from pants.backend.python.rules.prepare_chrooted_python_sources import ChrootedPythonSources
-
 from pants.backend.python.rules.python_test_runner import (
   DEFAULT_COVERAGE_CONFIG,
   get_coveragerc_input,
+  get_coverage_plugin_input,
+  COVERAGE_PLUGIN_MODULE_NAME,
+  construct_coverage_config,
 )
 from pants.backend.python.subsystems.pytest import PyTest
 from pants.backend.python.subsystems.python_setup import PythonSetup
@@ -40,7 +42,7 @@ from pants.engine.fs import (
 )
 from pants.engine.goal import Goal, GoalSubsystem
 from pants.engine.isolated_process import ExecuteProcessRequest, ExecuteProcessResult
-from pants.engine.legacy.graph import HydratedTarget, TransitiveHydratedTargets, HydratedTargets
+from pants.engine.legacy.graph import HydratedTarget, HydratedTargets, TransitiveHydratedTargets
 from pants.engine.rules import goal_rule, rule, subsystem_rule
 from pants.engine.selectors import Get, MultiGet
 from pants.rules.core.strip_source_root import SourceRootStrippedSources
@@ -155,51 +157,6 @@ async def merge_coverage_reports(
     request
   )
   return MergedCoverageData(coverage_data=result.output_directory_digest)
-
-
-COVERAGE_PLUGIN_MODULE_NAME = '__coverage_coverage_plugin__'
-
-
-def get_coverage_plugin_input():
-  return InputFilesContent(
-    FilesContent(
-      (
-        FileContent(
-          path=f'{COVERAGE_PLUGIN_MODULE_NAME}.py',
-          content=pkg_resources.resource_string(__name__, 'coverage/plugin.py'),
-          is_executable=False,
-        ),
-      )
-    )
-  )
-
-
-def ensure_section(config_parser: configparser.ConfigParser, section: str) -> None:
-  """Ensure a section exists in a ConfigParser."""
-  if not config_parser.has_section(section):
-    config_parser.add_section(section)
-
-
-def construct_coverage_config(source_roots, python_files) -> str:
-  # A map from source root stripped source to its source root. eg:
-  #  {'pants/testutil/subsystem/util.py': 'src/python'}
-  # This is so coverage reports referencing /chroot/path/pants/testutil/subsystem/util.py can be mapped
-  # back to the actual sources they reference when merging coverage reports.
-  source_to_target_base: Dict[str, str] = {}
-  for file_name in python_files:
-    source_root = source_roots.find_by_path(file_name)
-    source_root_stripped_path = file_name[len(source_root.path) + 1:]
-    source_to_target_base[source_root_stripped_path] = source_root.path
-
-  config_parser = configparser.ConfigParser()
-  config_parser.read_file(StringIO(DEFAULT_COVERAGE_CONFIG))
-  ensure_section(config_parser, 'run')
-  config_parser.set('run', 'plugins', COVERAGE_PLUGIN_MODULE_NAME)
-  config_parser.add_section(COVERAGE_PLUGIN_MODULE_NAME)
-  config_parser.set(COVERAGE_PLUGIN_MODULE_NAME, 'source_to_target_base', json.dumps(source_to_target_base))
-  config = StringIO()
-  config_parser.write(config)
-  return config.getvalue()
 
 
 def get_file_names(all_target_adaptors):
